+++ DelvUI/changelog.md
@@ -6,6 +6,7 @@ Features:
 - Added option to automatically disable the preview mode on HUD elements when the DelvUI's setting window is closed.
 - Added alpha bar to color selections.
 - Making party frames and status effects list areas more clear now. When you open DelvUI's settings window, these areas will automatically show reflecting the real size of the hud elements.
+- Added a cleanse tracker for party frames.
 - Added sub-option to Hide DelvUI outside of combat to show when weapons are drawn.
 - Added sub-option to Enable Combat Hotbars to always show in duties.
 - Added sub-option to Enable Combat Hotbars to show when weapons are drawn.